Reorganized files in gfx.
git-svn-id: http://svn.ta3d.org/yuni/trunk@1240 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/gfx/window/x11/opengl.h
@@ -32,7 +32,7 @@ namespace Window
 
 		virtual void close();
 
-		virtual void resize(size_t width, size_t height)
+		virtual void resize(unsigned int width, unsigned int height)
 		{
 			IWindow::resize(width, height);
 			Surface::OpenGL::resize(width, height);
